Moved the color tokens
